+++ packages/DS/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+- [x] Feat: change to purple as primary color for CTAs.
+
 ## 0.2.0
 
 - [x] Fix: UI / components improved.